[FIX, TST] Fix CircleCI (#148)
* Split pure Python and FSL-based IBMA tests.

* Update config file.

* Add requirements file.

* Use environment.

* Try restoring cache to get at environment.

* Use restore_cache.

* Update caching frequency.

* Make SCALE test less memory intensive.
