+++ nimare/tests/test_cbma_ale.py
@@ -2,7 +2,6 @@
 Test nimare.meta.cbma.ale (ALE/SCALE meta-analytic algorithms).
 """
 import pytest
-import numpy as np
 import pandas as pd
 
 import nimare
@@ -46,7 +45,7 @@ def test_ale(testdata1):
     Smoke test for ALE
     """
     ale_meta = ale.ALE(testdata1)
-    ale_meta.fit(n_iters=5, ids=testdata1.ids)
+    ale_meta.fit(n_iters=5, ids=testdata1.ids, n_cores=1)
     assert isinstance(ale_meta.results, nimare.base.meta.MetaResult)
 
 
@@ -55,15 +54,6 @@ def test_ale_subtraction(testdata1):
     Smoke test for ALE
     """
     ale_meta = ale.ALE(testdata1)
-    ale_meta.fit(n_iters=5, ids=testdata1.ids[:5], ids2=testdata1.ids[5:])
+    ale_meta.fit(n_iters=5, ids=testdata1.ids[:5], ids2=testdata1.ids[5:],
+                 n_cores=1)
     assert isinstance(ale_meta.results, nimare.base.meta.MetaResult)
-
-
-def test_scale(testdata1):
-    """
-    Smoke test for SCALE
-    """
-    ijk = np.vstack(np.where(testdata1.mask.get_data())).T
-    scale_meta = ale.SCALE(testdata1, ijk=ijk)
-    scale_meta.fit(n_iters=5, ids=testdata1.ids)
-    assert isinstance(scale_meta.results, nimare.base.meta.MetaResult)
