+++ requirements.txt
@@ -1,0 +1,17 @@
+nibabel
+numpy
+scipy
+pandas
+statsmodels
+nipype
+scikit-learn
+nilearn
+duecredit
+pyneurovault
+six
+matplotlib
+nltk
+fuzzywuzzy
+cognitiveatlas
+click
+tqdm
